make kubeapiserver collector more generic
